requirements.txt
PyExifTool==0.5.6
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
